+++ ParaTaskLambda/src/pt/runtime/TaskpoolLIFOWorkStealing.java
@@ -158,11 +158,9 @@ public class TaskpoolLIFOWorkStealing extends AbstractTaskPool {
 				taskinfo.setTaskInfoOfMultiTask(true);
 				
 				for (int i = 0; i < count; i++) {
-					TaskID<?> taskID = new TaskID(taskinfo);
+					TaskID taskID = new TaskID<>(taskinfo);
 					taskID.setRelativeID(i);
 					taskID.setExecuteOnThread(i%currentMultiTaskThreadPoolSize);
-					taskID.setSubTask(true);
-					taskID.setPartOfGroup(((TaskIDGroup<?>)nextTaskID));
 					((TaskIDGroup<?>)nextTaskID).addInnerTask(taskID);
 					enqueueReadyTask(taskID);
 				}